Create AppCompositionRoot by extracting common logic from activities and refactor application class
