+++ app/src/main/java/com/techyourchance/dagger2course/MyApplication.kt
@@ -1,6 +1,7 @@
 package com.techyourchance.dagger2course
 
 import android.app.Application
+import com.techyourchance.dagger2course.common.composition.AppCompositionRoot
 import com.techyourchance.dagger2course.networking.StackoverflowApi
 import com.techyourchance.dagger2course.questions.FetchQuestionDetailUseCase
 import com.techyourchance.dagger2course.questions.FetchQuestionUseCase
@@ -10,19 +11,9 @@ import retrofit2.create
 
 class MyApplication: Application() {
 
-    private val retrofit: Retrofit = Retrofit.Builder()
-        .baseUrl(Constants.BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    val stackoverflowApi: StackoverflowApi = retrofit.create(StackoverflowApi::class.java)
-
-    //get a new instance of fetchQuestion useCase when it is query, to use getter with get()
-    val fetchQuestionsUseCase get() = FetchQuestionUseCase(stackoverflowApi)
-
-    val fetchQuestionDetailUseCase get() = FetchQuestionDetailUseCase(stackoverflowApi)
-
+    lateinit var appCompositionRoot: AppCompositionRoot
     override fun onCreate() {
+        appCompositionRoot = AppCompositionRoot()
         super.onCreate()
     }
