+++ app/src/main/java/com/techyourchance/dagger2course/screens/questionslist/QuestionsListActivity.kt
@@ -7,12 +7,13 @@ import com.techyourchance.dagger2course.MyApplication
 import com.techyourchance.dagger2course.questions.FetchQuestionUseCase
 import com.techyourchance.dagger2course.questions.Question
 import com.techyourchance.dagger2course.screens.common.ScreensNavigator
+import com.techyourchance.dagger2course.screens.common.activities.BaseActivity
 import com.techyourchance.dagger2course.screens.common.dialogs.DialogsNavigator
 import com.techyourchance.dagger2course.screens.common.dialogs.ServerErrorDialogFragment
 import com.techyourchance.dagger2course.screens.questiondetails.QuestionDetailsActivity
 import kotlinx.coroutines.*
 
-class QuestionsListActivity : AppCompatActivity(),
+class QuestionsListActivity : BaseActivity(),
     QuestionListViewMvc.Listener {
 
     private val coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main.immediate)
@@ -34,7 +35,7 @@ class QuestionsListActivity : AppCompatActivity(),
 
         setContentView(viewMvc.rootView)
 
-        fetchQuestionUseCase = (application as MyApplication).fetchQuestionsUseCase
+        fetchQuestionUseCase = compositionRoot.fetchQuestionsUseCase
 
         dialogsNavigator = DialogsNavigator(supportFragmentManager)
